Set up GTP bot to use AlphaBeta
It evaluates every candidate at depth d, which is probably better described as
depth d+1.  With depth=2 (really 3), it plays in a couple of seconds on 9x9, but
not really well enough to tell if it is doing anything smart.  Slower than I
expected, but definitely lots of places the algorithm could be improved.
Wondering if game result evaluation (due to consideration of pass/resign) is a
significant factor.