Switch loss functions to means over samples
Use MSE for value and mean of cross-entropy loss.  This is I believe consistent
with what is shown in the papers, where they give the equation for a single
sample.

Also increasing the learning rate to 1e-2 as given in the papers, as it seems to
behave OK with this scaling of the loss.