Fix a bug in setting value for terminal node
The local value variable was shadowing the member variable, so the member value
was not actually updated for terminal nodes.  The effect was that the model's
predicted value was being used instead of replacing with the actual game result
value.  This probably had a significant adverse affect for endgame play and
passing.