Fix bug tracking move count across games
Make num_moves a member of GameState and use this to keep track for the zero
agent, as opposed to having its own value for this.  Previously, if multiple
games were played with the same agent instance, as in zero_sim.cpp, then the
agent's move count was not reset after each game.  This led to the initial move
randomization only taking place in the first game, so most training would not
have seen this move randomization.

This change also reinterprets the count to be total moves for both sides.  Not
100% clear from the AGZ paper, which says "For the first 30 moves of each
game..."