Move definition of hasher from header to cpp
This appears to fix the multiple initializations that I had previously seen when
activating the debug output.  Believe the reason is that when declared in the
header, each translation unit gets its own private copy of the global var.  Now
there is just one copy.  If it is also needed in the header, could add an extern
declaration.